FD-2462: rename video to vimeo
